+++ src/shape-component.js
@@ -64,6 +64,37 @@ class ShapeComponent {
   }
 
   /**
+   * @template T
+   * @param {string} name
+   * @param {T | (() => T)} value
+   * @param {any[]} [dependencies]
+   * @returns {T}
+   */
+  cacheStatic(name, value, dependencies) {
+    let actualValue
+    const constructor = /** @type {typeof ShapeComponent} */ (this.constructor)
+
+    if (!constructor.__staticCaches) {
+      constructor.__staticCaches = {}
+    }
+
+    const oldDependencies = constructor.__staticCaches[name]?.dependencies
+
+    if (typeof value == "function") {
+      // @ts-expect-error
+      actualValue = value()
+    } else {
+      actualValue = value
+    }
+
+    if (!(name in constructor.__staticCaches) || arrayReferenceDifferent(oldDependencies || [], dependencies || [])) {
+      constructor.__staticCaches[name] = {dependencies, value: actualValue}
+    }
+
+    return constructor.__staticCaches[name].value
+  }
+
+  /**
    * @param {Record<string, any>} variables
    * @returns {void}
    */